+++ api/mocks/mock_full.go
@@ -1492,36 +1492,6 @@ func (mr *MockFullNodeMockRecorder) EthSyncing(arg0 interface{}) *gomock.Call {
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "EthSyncing", reflect.TypeOf((*MockFullNode)(nil).EthSyncing), arg0)
 }
 
-// EthTraceBlock mocks base method.
-func (m *MockFullNode) EthTraceBlock(arg0 context.Context, arg1 string) ([]*ethtypes.EthTraceBlock, error) {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "EthTraceBlock", arg0, arg1)
-	ret0, _ := ret[0].([]*ethtypes.EthTraceBlock)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// EthTraceBlock indicates an expected call of EthTraceBlock.
-func (mr *MockFullNodeMockRecorder) EthTraceBlock(arg0, arg1 interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "EthTraceBlock", reflect.TypeOf((*MockFullNode)(nil).EthTraceBlock), arg0, arg1)
-}
-
-// EthTraceReplayBlockTransactions mocks base method.
-func (m *MockFullNode) EthTraceReplayBlockTransactions(arg0 context.Context, arg1 string, arg2 []string) ([]*ethtypes.EthTraceReplayBlockTransaction, error) {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "EthTraceReplayBlockTransactions", arg0, arg1, arg2)
-	ret0, _ := ret[0].([]*ethtypes.EthTraceReplayBlockTransaction)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// EthTraceReplayBlockTransactions indicates an expected call of EthTraceReplayBlockTransactions.
-func (mr *MockFullNodeMockRecorder) EthTraceReplayBlockTransactions(arg0, arg1, arg2 interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "EthTraceReplayBlockTransactions", reflect.TypeOf((*MockFullNode)(nil).EthTraceReplayBlockTransactions), arg0, arg1, arg2)
-}
-
 // EthUninstallFilter mocks base method.
 func (m *MockFullNode) EthUninstallFilter(arg0 context.Context, arg1 ethtypes.EthFilterID) (bool, error) {
 	m.ctrl.T.Helper()
