+++ documentation/en/api-v1-unstable-methods.md
@@ -104,8 +104,6 @@
   * [EthSendRawTransaction](#EthSendRawTransaction)
   * [EthSubscribe](#EthSubscribe)
   * [EthSyncing](#EthSyncing)
-  * [EthTraceBlock](#EthTraceBlock)
-  * [EthTraceReplayBlockTransactions](#EthTraceReplayBlockTransactions)
   * [EthUninstallFilter](#EthUninstallFilter)
   * [EthUnsubscribe](#EthUnsubscribe)
 * [Filecoin](#Filecoin)
@@ -3074,99 +3072,6 @@ Perms: read
 Inputs: `null`
 
 Response: `false`
-
-### EthTraceBlock
-TraceAPI related methods
-
-Returns traces created at given block
-
-
-Perms: read
-
-Inputs:
-```json
-[
-  "string value"
-]
-```
-
-Response:
-```json
-[
-  {
-    "action": {
-      "callType": "string value",
-      "from": "0x5cbeecf99d3fdb3f25e309cc264f240bb0664031",
-      "to": "0x5cbeecf99d3fdb3f25e309cc264f240bb0664031",
-      "gas": "0x5",
-      "input": "0x07",
-      "value": "0x0"
-    },
-    "result": {
-      "gasUsed": "0x5",
-      "output": "0x07"
-    },
-    "subtraces": 123,
-    "traceAddress": [
-      123
-    ],
-    "Type": "string value",
-    "blockHash": "0x37690cfec6c1bf4c3b9288c7a5d783e98731e90b0a4c177c2a374c7a9427355e",
-    "blockNumber": 9,
-    "transactionHash": "0x37690cfec6c1bf4c3b9288c7a5d783e98731e90b0a4c177c2a374c7a9427355e",
-    "transactionPosition": 123
-  }
-]
-```
-
-### EthTraceReplayBlockTransactions
-Replays all transactions in a block returning the requested traces for each transaction
-
-
-Perms: read
-
-Inputs:
-```json
-[
-  "string value",
-  [
-    "string value"
-  ]
-]
-```
-
-Response:
-```json
-[
-  {
-    "output": "0x07",
-    "stateDiff": "string value",
-    "trace": [
-      {
-        "action": {
-          "callType": "string value",
-          "from": "0x5cbeecf99d3fdb3f25e309cc264f240bb0664031",
-          "to": "0x5cbeecf99d3fdb3f25e309cc264f240bb0664031",
-          "gas": "0x5",
-          "input": "0x07",
-          "value": "0x0"
-        },
-        "result": {
-          "gasUsed": "0x5",
-          "output": "0x07"
-        },
-        "subtraces": 123,
-        "traceAddress": [
-          123
-        ],
-        "Type": "string value"
-      }
-    ],
-    "transactionHash": "0x37690cfec6c1bf4c3b9288c7a5d783e98731e90b0a4c177c2a374c7a9427355e",
-    "vmTrace": "string value"
-  }
-]
-```
 
 ### EthUninstallFilter
 Uninstalls a filter with given id.
